Mount devtmpfs read-only into chroot
It's too easy to destroy the content of devtmpfs, which is shared with
the host (including privileged container setups), by calling rm -rf on
an output dir that still has devtmpfs mounted.

To achieve write protection for device nodes, we can't mount devtmpfs
directly in read-only mode as that will change all mounts to that mode.
Luckily, doing a read-only bind-mount does the trick.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>